+++ docs/changelog.md
@@ -5,3 +5,8 @@ Change Log
 *2015 Mar. 06*
 
 1. client-app-driver now changes document.title after excuting Navigate action.
+
+##v0.4.3
+1. RequireFilter is now a class
+2. Requirefilter would return url friendly paths with versions
+3. Cache buster is built into app-loader's initial scripts